refactoring in progress 6
